requirements.txt
biopython==1.84
streamlit==1.39.0